refactor: migrate benchmarks to ESM modules
